# Optimize unsafe_sample