In `unsafe_sample` the `sample_times` is guaranteed to be a numpy array. Calling `searchsorted` from the array is more efficient.

For a small numpy array:
```
In [66]: %timeit sample_times.searchsorted( (0,0), 'left')
1.53 µs ± 63.9 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [67]: %timeit np.searchsorted(sample_times, (0,0), 'left')
2.73 µs ± 424 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

@terrorfisch