@terrorfisch 2023-08-24
Thanks. Seems like numpy's `array_function_dispatch` introduces quite some overhead.